feat: merge radio page into toggle
