+++ src/components/unity-hierarchy/toggle.tsx
@@ -54,10 +54,35 @@ const ToggleGroup = (props: Partial<TreeItemProps> = {}): ComponentResult => {
   };
 };
 
+const RadioToggle = (props: Partial<TreeItemProps> = {}): ComponentResult => {
+  const TextH5 = Text.H5();
+  const displayName = props.name ?? "Radio";
+
+  return {
+    data: {
+      title: "Radio",
+      img: "https://via.placeholder.com/500x400",
+      href: "/vrc/monoui/radio",
+      assetsPath: "/Packages/Mono UI/Runtime/Assets/Prefab/Components/Radio.prefab",
+      contextMenuPath: "GameObject/Mono UI/Toggle/Radio Toggle",
+      dependencies: [TextH5.data],
+    },
+    content: (
+      <Tree.Item name={displayName} type={Package} {...props}>
+        <Tree.Item name="Background" type={Box}>
+          <Tree.Item name="Checkmark" type={Box} />
+        </Tree.Item>
+        {TextH5.content}
+      </Tree.Item>
+    ),
+  };
+};
+
 const Toggle: Record<string, (props?: Partial<TreeItemProps>) => ComponentResult> = {
   Default: ToggleDefault,
   Outline: (props) => ToggleDefault({ name: "Toggle Outline", ...props }),
   Group: ToggleGroup,
+  Radio: RadioToggle,
 };
 
 export default Toggle;
