--- src/pages/vrc/monoui/radio.mdx
@@ -1,33 +1,0 @@
-import Image from "next/image";
-import { Tabs, Cards } from "nextra/components";
-import { Box, Package } from "lucide-react";
-import Tree from "../../../components/tree.tsx";
-import ComponentsTabs from "../../../components/components-tabs.tsx";
-import Radio from "../../../components/unity-hierarchy/radio.tsx";
-
-# Radio
-
-Displays a radio button or a component that looks like a radio button.
-
-<ComponentsTabs component={Radio.Default()} />
-
-## Installation
-
-Not required.
-
-## Usage
-
-<Tree.Container>
-  <Tree.Root defaultExpanded>
-    <Tree.Item name="Canvas" type={Box}>
-      {Radio.Default({ defaultExpanded: false }).content}
-      {Radio.Default({ defaultExpanded: false }).content}
-    </Tree.Item>
-  </Tree.Root>
-</Tree.Container>
-
-## Examples
-
-### Primary
-
-<ComponentsTabs component={Radio.Default()} />
