+++ src/pages/vrc/monoui/toggle.mdx
@@ -39,3 +39,7 @@ Not required.
 ### Group
 
 <ComponentsTabs component={Toggle.Group()} />
+
+### Radio
+
+<ComponentsTabs component={Toggle.Radio()} />
